modularizado el primer endpoint obtener todos los juegos y juegos por nombres
