+++ api/src/routes/routes.js
@@ -1,8 +1,8 @@
 require('dotenv').config();
 const { Router } = require('express');
 const axios = require('axios')
-const { Op } = require('Sequelize')
 const { Videogame, Genre } = require('../db')
+const allgames = require('../controllers/getAllvideogames.js')
 const { API_KEY } = process.env
 // Importar todos los routers;
 // Ejemplo: const authRouter = require('./auth.js');
@@ -12,48 +12,16 @@ const router = Router();
 
 // Configurar los routers
 // Ejemplo: router.use('/auth', authRouter);
-const ENDPAPI1 = 'https://api.rawg.io/api/games?key='
-const ENDPAPI2 = 'https://api.rawg.io/api/games?search='
+
+
 const ENDPAPI3 = 'https://api.rawg.io/api/genres'
 const ENDPAPI4 = 'https://api.rawg.io/api/games/'
 const ENDPAPI5 = 'https://api.rawg.io/api/platforms'
 
 
-let allgames = []
-async function traertodo() {
-    if (allgames.length > 0) return allgames
-    let page2, page3, page4, page5
-    let { data } = await axios.get(`${ENDPAPI1}${API_KEY}`)
-    let page1R = data.results //+ 20 juegos
-    let page1D = reducedata(page1R)
-    page2 = await axios.get(`${data.next}`)
-    let page2R = page2.data.results//+ 20 juegos 
-    let page2D = reducedata(page2R)
-    page3 = await axios.get(`${page2.data.next}`)
-    let page3R = page3.data.results //+ 20 juegos 
-    let page3D = reducedata(page3R)
-    page4 = await axios.get(`${page3.data.next}`)
-    let page4R = page4.data.results //+ 20 juegos 
-    let page4D = reducedata(page4R)
-    page5 = await axios.get(`${page4.data.next}`)
-    let page5R = page5.data.results //+ 20 juegos */
-    let page5D = reducedata(page5R)
-    allgames = [...page1D, ...page2D, ...page3D, ...page4D, ...page5D]
-    return allgames // retorna el arreglo de todos los juegos con los valores que queremos
-}
 
-function reducedata(array) {
-    return array.map(Obj => {
-        return {
-            id: Obj.id,
-            image: Obj.background_image || Obj.image,
-            name: Obj.name,
-            genres: Obj.genres.map(g => g.name),
-            rating: Obj.rating,
-            CreatedInDb: Obj.CreatedInDb ? Obj.CreatedInDb : false
-        }
-    })
-}
+
+
 
 function reducebyid(Obj) {
     return {
@@ -95,93 +63,16 @@ async function reducePlatfomr() {
     }
 }
 
-function orderbyname(orden, all) {
-    const nameorder = orden === 'asd' ? all.sort(function (a, b) {
-        const onename = a.name.toLowerCase()
-        const twoname = b.name.toLowerCase()
-        if (onename > twoname) {
-            return 1;
-        }
-        if (onename < twoname) {
-            return -1;
-        }
 
-        return 0;
-    }) : all.sort(function (a, b) {
-        const onename = a.name.toLowerCase()
-        const twoname = b.name.toLowerCase()
-        if (onename < twoname) {
-            return 1;
-        }
-        if (onename > twoname) {
-            return -1;
-        }
-        return 0;
-    })
-    return nameorder
-}
 
-function filter(filt, all) {
-    const allVideogames = all
-    const genresfilter = filt === 'ALL' ? allVideogames : allVideogames.filter(el => el.genres.includes(filt))
-    return genresfilter
-}
 
 const update = async (videogame, genres) => {
 
 }
 
-router.get('/videogames', async (req, res) => {
-    // const { name } = req.params
-    const { name, order, filtro } = req.query
-    if (!name) {
-        const allrequestApi = await traertodo()
-        const allrequestBd = await Videogame.findAll({
-            include: {
-                model: Genre,
-                attributes: ['name'],
+router.get('/videogames', allgames)
 
-                through: {
-                    attributes: []
-                }
-            }
-        })
 
-        const allrequest = [...reducedata(allrequestBd), ...allrequestApi]
-        filtro && order ? res.json(filter(filtro, orderbyname(order, allrequest)))
-            : order ? res.json(orderbyname(order, allrequest))
-                : filtro ? res.json(filter(filtro, allrequest))
-                    : res.json(allrequest)
-
-    } else {
-        try {
-            const { data } = await axios.get(`${ENDPAPI2}${name}&key=${API_KEY}`)
-            const resultadosApi = data.results
-            const nameBd = await Videogame.findAll({
-                where: { name: { [Op.iLike]: `%${name}%` } },
-                include: {
-                    model: Genre,
-                    attributes: ['name'],
-                    through: {
-                        attributes: []
-                    }
-                }
-            })
-            if (resultadosApi.length > 0 || nameBd.length > 0) {
-                const results = 15 - nameBd.length
-                const resto = [...reducedata(nameBd), ...reducedata(resultadosApi).slice(0, results)]
-                order && filtro ? res.json(filter(filtro, orderbyname(order, resto))) :
-                    order ? res.json(orderbyname(order, resto)) :
-                        filtro ? res.json(filter(filtro, resto))
-                            : res.send(resto)
-            } else {
-                return res.status(404).send('no encontrado')
-            }
-        } catch (error) {
-            res.send(error)
-        }
-    }
-})
 
 router.get('/videogames/:id', async (req, res) => {
     const { id } = req.params
@@ -264,7 +155,8 @@ router.put('/update/:id', async (req, res) => {
     try {
         let params = {}
         for (let value in req.body) {
-            if (value !== 'genresid') {1
+            if (value !== 'genresid') {
+                1
                 params[value] = req.body[value]
             }
         }
